+++ apps/melding-form/src/app/(map)/locatie/kies/_components/Map/components/Notification/Notification.tsx
@@ -6,7 +6,7 @@ import styles from './Notification.module.css'
 export type Props = Omit<AlertProps, 'headingLevel'> & { description?: string }
 
 export const Notification = ({ description, ...restProps }: Props) => (
-  <Alert {...restProps} closeable className={styles.notification} headingLevel={2} tabIndex={-1} aria-live="polite">
+  <Alert {...restProps} closeable className={styles.notification} headingLevel={2} tabIndex={-1} aria-live="assertive">
     {description && <Paragraph>{description}</Paragraph>}
   </Alert>
 )
