+++ apps/melding-form/src/app/(map)/locatie/kies/hooks/useAssetLayer.ts
@@ -5,7 +5,8 @@ import { Dispatch, SetStateAction, useEffect, useRef } from 'react'
 import { Feature } from '@meldingen/api-client'
 
 import { addAssetLayerToMap, fetchAssets, updateAssetMarkers } from '../_utils'
-import { Coordinates, NotificationType } from 'apps/melding-form/src/types'
+import { NotificationType } from '../types'
+import { Coordinates } from 'apps/melding-form/src/types'
 
 export type Props = {
   assetList: Feature[]
