+++ apps/melding-form/src/app/(map)/locatie/kies/types.d.ts
@@ -1,0 +1,8 @@
+import type { AlertProps } from '@amsterdam/design-system-react'
+
+export type NotificationType = {
+  closeButtonLabel: AlertProps['closeButtonLabel']
+  description?: string
+  heading: AlertProps['heading']
+  severity?: AlertProps['severity']
+}
