+++ apps/melding-form/src/types.d.ts
@@ -1,5 +1,3 @@
-import type { AlertProps } from '@amsterdam/design-system-react'
-
 export type Coordinates = { lat: number; lng: number }
 
 export type ValidationError = {
@@ -11,11 +9,4 @@ export type FormState = {
   formData?: FormData
   systemError?: unknown
   validationErrors?: ValidationError[]
-}
-
-export type NotificationType = {
-  closeButtonLabel: AlertProps['closeButtonLabel']
-  description?: string
-  heading: AlertProps['heading']
-  severity?: AlertProps['severity']
 }
